app is kinda functioning, we are back to the loggin bug, but create works
