+++ app/javascript/components/pages/ProtectedStudentIndex.js
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React from "react";
 import {
   Card,
   CardBody,
@@ -10,23 +10,10 @@ import {
 } from "reactstrap";
 import { NavLink } from "react-router-dom";
 
-const ProtectedStudentIndex = ({}) => {
+const ProtectedStudentIndex = ({students}) => {
   
-  const [students, setStudents] = useState([]);
 
-  useEffect(() => {
-    readStudents()
-  }, [])
 
-  const readStudents = () => {
-    fetch("/students")
-      .then((response) => response.json())
-      .then((payload) => {
-        console.log(payload)
-        setStudents(payload)
-      })
-      .catch((error) => console.log(error))
-  }
   return (
     <>
       <h1>View Your Students Here.</h1>
